fix: fixing bug where tray menu click does not trigger.

+++ crates/niva/src/app/api/window_extra.rs
@@ -57,7 +57,7 @@ pub fn register_api_instances(api_manager: &mut ApiManager) {
 // windows
 #[cfg(target_os = "windows")]
 #[niva_api]
-fn set_enable(enabled: bool, id: Option<u16>) -> Result<()> {
+fn set_enable(enabled: bool, id: Option<u8>) -> Result<()> {
     match_window!(app, window, id);
     window.set_enable(enabled);
     Ok(())
@@ -65,7 +65,7 @@ fn set_enable(enabled: bool, id: Option<u16>) -> Result<()> {
 
 #[cfg(target_os = "windows")]
 #[niva_api]
-fn set_taskbar_icon(taskbar_icon: String, id: Option<u16>) -> Result<()> {
+fn set_taskbar_icon(taskbar_icon: String, id: Option<u8>) -> Result<()> {
     match_window!(app, window, id);
     let taskbar_icon = app.resource().load_icon(&taskbar_icon)?;
     window.set_taskbar_icon(Some(taskbar_icon));
@@ -74,7 +74,7 @@ fn set_taskbar_icon(taskbar_icon: String, id: Option<u16>) -> Result<()> {
 
 #[cfg(target_os = "windows")]
 #[niva_api]
-fn theme(id: Option<u16>) -> Result<String> {
+fn theme(id: Option<u8>) -> Result<String> {
     match_window!(app, window, id);
     match window.theme() {
         Theme::Dark => Ok("dark".to_string()),
@@ -85,7 +85,7 @@ fn theme(id: Option<u16>) -> Result<String> {
 
 #[cfg(target_os = "windows")]
 #[niva_api]
-fn reset_dead_keys(id: Option<u16>) -> Result<()> {
+fn reset_dead_keys(id: Option<u8>) -> Result<()> {
     match_window!(app, window, id);
     window.reset_dead_keys();
     Ok(())
@@ -93,7 +93,7 @@ fn reset_dead_keys(id: Option<u16>) -> Result<()> {
 
 #[cfg(target_os = "windows")]
 #[niva_api]
-fn begin_resize_drag(edge: isize, button: u32, x: i32, y: i32, id: Option<u16>) -> Result<()> {
+fn begin_resize_drag(edge: isize, button: u32, x: i32, y: i32, id: Option<u8>) -> Result<()> {
     match_window!(app, window, id);
     window.begin_resize_drag(edge, button, x, y);
     Ok(())
@@ -101,7 +101,7 @@ fn begin_resize_drag(edge: isize, button: u32, x: i32, y: i32, id: Option<u16>) 
 
 #[cfg(target_os = "windows")]
 #[niva_api]
-fn set_skip_taskbar(skip: bool, id: Option<u16>) -> Result<()> {
+fn set_skip_taskbar(skip: bool, id: Option<u8>) -> Result<()> {
     match_window!(app, window, id);
     window.set_skip_taskbar(skip);
     Ok(())
@@ -109,7 +109,7 @@ fn set_skip_taskbar(skip: bool, id: Option<u16>) -> Result<()> {
 
 #[cfg(target_os = "windows")]
 #[niva_api]
-fn set_undecorated_shadow(shadow: bool, id: Option<u16>) -> Result<()> {
+fn set_undecorated_shadow(shadow: bool, id: Option<u8>) -> Result<()> {
     match_window!(app, window, id);
     window.set_undecorated_shadow(shadow);
     Ok(())
@@ -117,28 +117,28 @@ fn set_undecorated_shadow(shadow: bool, id: Option<u16>) -> Result<()> {
 
 #[cfg(target_os = "macos")]
 #[niva_api]
-fn simple_fullscreen(id: Option<u16>) -> Result<bool> {
+fn simple_fullscreen(id: Option<u8>) -> Result<bool> {
     match_window!(app, window, id);
     Ok(window.simple_fullscreen())
 }
 
 #[cfg(target_os = "macos")]
 #[niva_api]
-fn set_simple_fullscreen(fullscreen: bool, id: Option<u16>) -> Result<bool> {
+fn set_simple_fullscreen(fullscreen: bool, id: Option<u8>) -> Result<bool> {
     match_window!(app, window, id);
     Ok(window.set_simple_fullscreen(fullscreen))
 }
 
 #[niva_api]
 #[cfg(target_os = "macos")]
-fn has_shadow(id: Option<u16>) -> Result<bool> {
+fn has_shadow(id: Option<u8>) -> Result<bool> {
     match_window!(app, window, id);
     Ok(window.has_shadow())
 }
 
 #[cfg(target_os = "macos")]
 #[niva_api]
-fn set_has_shadow(has_shadow: bool, id: Option<u16>) -> Result<()> {
+fn set_has_shadow(has_shadow: bool, id: Option<u8>) -> Result<()> {
     match_window!(app, window, id);
     window.set_has_shadow(has_shadow);
     Ok(())
@@ -146,7 +146,7 @@ fn set_has_shadow(has_shadow: bool, id: Option<u16>) -> Result<()> {
 
 #[cfg(target_os = "macos")]
 #[niva_api]
-fn set_is_document_edited(edited: bool, id: Option<u16>) -> Result<()> {
+fn set_is_document_edited(edited: bool, id: Option<u8>) -> Result<()> {
     match_window!(app, window, id);
     window.set_is_document_edited(edited);
     Ok(())
@@ -154,14 +154,14 @@ fn set_is_document_edited(edited: bool, id: Option<u16>) -> Result<()> {
 
 #[cfg(target_os = "macos")]
 #[niva_api]
-fn is_document_edited(id: Option<u16>) -> Result<bool> {
+fn is_document_edited(id: Option<u8>) -> Result<bool> {
     match_window!(app, window, id);
     Ok(window.is_document_edited())
 }
 
 #[cfg(target_os = "macos")]
 #[niva_api]
-fn set_allows_automatic_window_tabbing(enabled: bool, id: Option<u16>) -> Result<()> {
+fn set_allows_automatic_window_tabbing(enabled: bool, id: Option<u8>) -> Result<()> {
     match_window!(app, window, id);
     window.set_allows_automatic_window_tabbing(enabled);
     Ok(())
@@ -169,14 +169,14 @@ fn set_allows_automatic_window_tabbing(enabled: bool, id: Option<u16>) -> Result
 
 #[cfg(target_os = "macos")]
 #[niva_api]
-fn allows_automatic_window_tabbing(id: Option<u16>) -> Result<bool> {
+fn allows_automatic_window_tabbing(id: Option<u8>) -> Result<bool> {
     match_window!(app, window, id);
     Ok(window.allows_automatic_window_tabbing())
 }
 
 #[cfg(target_os = "macos")]
 #[niva_api]
-fn set_tabbing_identifier(identifier: String, id: Option<u16>) -> Result<()> {
+fn set_tabbing_identifier(identifier: String, id: Option<u8>) -> Result<()> {
     match_window!(app, window, id);
     window.set_tabbing_identifier(&identifier);
     Ok(())
@@ -184,7 +184,7 @@ fn set_tabbing_identifier(identifier: String, id: Option<u16>) -> Result<()> {
 
 #[cfg(target_os = "macos")]
 #[niva_api]
-fn tabbing_identifier(id: Option<u16>) -> Result<String> {
+fn tabbing_identifier(id: Option<u8>) -> Result<String> {
     match_window!(app, window, id);
     Ok(window.tabbing_identifier())
 }
